+++ gulpfile.js/tasks/watch.js
@@ -26,8 +26,21 @@ gulp.task('watch', ['js-watch'], function (callback) {
     watch(config.source.getFileGlobs('svg'),
         function (events, done) { gulp.start('svg'); });
 
+    watch(config.source.getPath('components', '**/*.scss'),
+        function (events, done) {
+            gulp.start('css');
+            gulp.start('css-lint');
+            gulp.start('inject-component-css');
+        });
+
     watch(config.source.getPath('css', '**/*.scss'),
-        function (events, done) { gulp.start('css'); gulp.start('css-lint'); });
+        function (events, done) {
+            gulp.start('css');
+            gulp.start('css-lint');
+        });
+
+    watch(config.source.getPath('components', '**/*.html'),
+        function (events, done) { gulp.start('html'); });
 
     watch(config.source.getPath('html', '**'),
         function (events, done) { gulp.start('html'); });